e2e pings should work on mobile
