+++ react/features/base/conference/actionTypes.js
@@ -83,6 +83,20 @@ export const CONFERENCE_TIMESTAMP_CHANGED = 'CONFERENCE_TIMESTAMP_CHANGED';
 export const CONFERENCE_UNIQUE_ID_SET = 'CONFERENCE_UNIQUE_ID_SET';
 
 /**
+ * Create an action for when an end-to-end RTT is changed.
+ *
+ * {
+ *     type: E2E_RTT_CHANGED,
+ *     e2eRtt: {
+ *         rtt: number,
+ *         remoteEndpointId: string,
+ *         remoteRegion: string
+ *     }
+ * }
+ */
+export const E2E_RTT_CHANGED = 'E2E_RTT_CHANGED'
+
+/**
  * The type of (redux) action which signals that a specific conference will be
  * joined.
  *
