+++ react/features/base/conference/actions.js
@@ -9,7 +9,7 @@ import {
 import { endpointMessageReceived } from '../../subtitles';
 import { getReplaceParticipant } from '../config/functions';
 import { JITSI_CONNECTION_CONFERENCE_KEY } from '../connection';
-import { JitsiConferenceEvents } from '../lib-jitsi-meet';
+import {JitsiConferenceEvents, JitsiE2ePingEvents} from '../lib-jitsi-meet';
 import {
     MEDIA_TYPE,
     setAudioMuted,
@@ -48,6 +48,7 @@ import {
     CONFERENCE_WILL_JOIN,
     CONFERENCE_WILL_LEAVE,
     DATA_CHANNEL_OPENED,
+    E2E_RTT_CHANGED,
     KICKED_OUT,
     LOCK_STATE_CHANGED,
     NON_PARTICIPANT_MESSAGE_RECEIVED,
@@ -232,6 +233,10 @@ function _addConferenceListeners(conference, dispatch, state) {
         (...args) => dispatch(participantPresenceChanged(...args)));
 
     conference.on(
+        JitsiE2ePingEvents.E2E_RTT_CHANGED,
+        (...args) => dispatch(e2eRttChanged(...args)));
+
+    conference.on(
         JitsiConferenceEvents.BOT_TYPE_CHANGED,
         (id, botType) => dispatch(participantUpdated({
             conference,
@@ -253,6 +258,32 @@ function _addConferenceListeners(conference, dispatch, state) {
             id,
             email: data.value
         })));
+}
+
+
+/**
+ * Create an action for when a participant e2e rtt is changed.
+ *
+ * @param {Object} participant - The participant against which the rtt is measured.
+ * @param {number} rtt - The rtt.
+ * @returns {{
+ *     type: E2E_RTT_CHANGED,
+ *     e2eRtt: {
+ *         remoteEndpointId: string,
+ *         remoteRegion: string,
+ *         rtt: number
+ *     }
+ * }}
+ */
+export function e2eRttChanged(participant, rtt) {
+    return {
+        type: E2E_RTT_CHANGED,
+        e2eRtt: {
+            rtt,
+            remoteEndpointId: participant.getId(),
+            remoteRegion: participant.getProperty('region')
+        }
+    };
 }
 
 /**
