+++ react/features/rtcstats/middleware.js
@@ -3,12 +3,9 @@
 import { jitsiLocalStorage } from '@jitsi/js-utils';
 
 import { getAmplitudeIdentity } from '../analytics';
-import { CONFERENCE_UNIQUE_ID_SET, getConferenceOptions, getRoomName } from '../base/conference';
+import { CONFERENCE_UNIQUE_ID_SET, E2E_RTT_CHANGED, getConferenceOptions, getRoomName } from '../base/conference';
 import { LIB_WILL_INIT } from '../base/lib-jitsi-meet';
-import {
-    DOMINANT_SPEAKER_CHANGED, E2E_RTT_CHANGED,
-    getLocalParticipant
-} from '../base/participants';
+import { DOMINANT_SPEAKER_CHANGED, getLocalParticipant } from '../base/participants';
 import { MiddlewareRegistry } from '../base/redux';
 import { ADD_FACIAL_EXPRESSION } from '../facial-recognition/actionTypes';
